Rename file so it does not look like a test image
