+++ napari/_qt/experimental/render/image_creator.py
